# Conflicts:
#	app/src/main/java/com/example/studentintelligentsystem/ManageStudentsActivity.java
#	app/src/main/res/drawable/ic_admin.xml
#	app/src/main/res/layout/activity_manage_students.xml

+++ app/src/main/java/com/example/studentintelligentsystem/DatabaseHelper.java
@@ -9,7 +9,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String DATABASE_NAME = "student_system.db";
-    private static final int DATABASE_VERSION = 2;
+    private static final int DATABASE_VERSION = 3;
 
     // Tables
     private static final String TABLE_STUDENTS = "students";
@@ -22,6 +22,13 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     private static final String COL_STU_ID = "id";
     private static final String COL_STU_NAME = "stu_name";
     private static final String COL_STU_GRADE = "stu_grade";
+    private static final String COL_STU_AGE = "stu_age";
+    private static final String COL_STU_GENDER = "stu_gender";
+    private static final String COL_STU_PARENT_NAME = "stu_parent_name";
+    private static final String COL_STU_PARENT_EMAIL = "stu_parent_email";
+    private static final String COL_STU_PARENT_PHONE = "stu_parent_phone";
+    private static final String COL_STU_ADDRESS = "stu_address";
+
     // Results columns
     private static final String COL_RES_ID = "id";
     private static final String COL_RES_STU_ID = "student_id";
@@ -53,7 +60,13 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         String createStudentsTable = "CREATE TABLE " + TABLE_STUDENTS + " (" +
                 COL_STU_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 COL_STU_NAME + " TEXT, " +
-                COL_STU_GRADE + " INTEGER)";
+                COL_STU_GRADE + " INTEGER, " +
+                COL_STU_AGE + " INTEGER, " +
+                COL_STU_GENDER + " TEXT, " +
+                COL_STU_PARENT_NAME + " TEXT, " +
+                COL_STU_PARENT_EMAIL + " TEXT, " +
+                COL_STU_PARENT_PHONE + " TEXT, " +
+                COL_STU_ADDRESS + " TEXT)";
         db.execSQL(createStudentsTable);
 
         // Create parents table
@@ -90,17 +103,42 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // Simple upgrade: create new tables if they don't exist
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_RESULTS);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_ATTENDANCE);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_STUDENTS);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_PARENTS);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
-        onCreate(db);
+        if (oldVersion < 2) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_RESULTS);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_ATTENDANCE);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_STUDENTS);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_PARENTS);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
+            onCreate(db);
+        }
+
+        if (oldVersion < 3) {
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_AGE + " INTEGER");
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_GENDER + " TEXT");
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_PARENT_NAME + " TEXT");
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_PARENT_EMAIL + " TEXT");
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_PARENT_PHONE + " TEXT");
+            db.execSQL("ALTER TABLE " + TABLE_STUDENTS + " ADD COLUMN " + COL_STU_ADDRESS + " TEXT");
+        }
     }
 
 
     // STUDENT METHODS
+
+    public boolean addStudent(Student student) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        ContentValues cv = new ContentValues();
+        cv.put(COL_STU_NAME, student.getName());
+        cv.put(COL_STU_GRADE, student.getGrade());
+        cv.put(COL_STU_AGE, student.getAge());
+        cv.put(COL_STU_GENDER, student.getGender());
+        cv.put(COL_STU_PARENT_NAME, student.getParentName());
+        cv.put(COL_STU_PARENT_EMAIL, student.getParentEmail());
+        cv.put(COL_STU_PARENT_PHONE, student.getParentPhone());
+        cv.put(COL_STU_ADDRESS, student.getAddress());
+        long result = db.insert(TABLE_STUDENTS, null, cv);
+        return result != -1;
+    }
 
     public boolean addStudent(String name, int grade) {
         SQLiteDatabase db = this.getWritableDatabase();
@@ -109,6 +147,21 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         cv.put(COL_STU_GRADE, grade);
         long result = db.insert(TABLE_STUDENTS, null, cv);
         return result != -1;
+    }
+
+    public boolean updateStudent(Student student) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        ContentValues cv = new ContentValues();
+        cv.put(COL_STU_NAME, student.getName());
+        cv.put(COL_STU_GRADE, student.getGrade());
+        cv.put(COL_STU_AGE, student.getAge());
+        cv.put(COL_STU_GENDER, student.getGender());
+        cv.put(COL_STU_PARENT_NAME, student.getParentName());
+        cv.put(COL_STU_PARENT_EMAIL, student.getParentEmail());
+        cv.put(COL_STU_PARENT_PHONE, student.getParentPhone());
+        cv.put(COL_STU_ADDRESS, student.getAddress());
+        int rows = db.update(TABLE_STUDENTS, cv, COL_STU_ID + "=?", new String[]{String.valueOf(student.getId())});
+        return rows > 0;
     }
 
     public boolean updateStudent(int id, String name, int grade) {
